# Conflicts:
#	pyttman/core/internals.py

+++ tests/tools/logger/test_logger.py
@@ -1,4 +1,3 @@
-
 import logging
 import os
 import sys
@@ -8,12 +7,9 @@ import pyttman
 from tests.module_helper import PyttmanInternalBaseTestCase
 
 
-@pyttman.logger.loggedmethod
-def some_func():
-    raise Exception("This is a log message")
-
-
 class TestPyttmanLogger(PyttmanInternalBaseTestCase):
+
+    cleanup_after = False
 
     def test_logger_as_class(self):
         expected_output_in_file = "DEBUG:PyttmanTestLogger:This is a log message"
@@ -21,10 +17,23 @@ class TestPyttmanLogger(PyttmanInternalBaseTestCase):
         self.logfile_meets_expectation(expected_output_in_file)
 
     def test_logger_as_decorator(self):
-        expected_output_in_file = 'raise Exception("This is a log message")'
+        @pyttman.logger
+        def broken():
+            raise Exception("This is a log message")
 
-        self.assertRaises(Exception, some_func)
+        @pyttman.logger()
+        def working():
+            return "I work"
+
+        working()
         self.assertTrue(Path(self.log_file_name).exists())
+
+        expected_output_in_file = "Return value from 'working': 'I work'"
+        self.logfile_meets_expectation(expected_output_in_file)
+
+        with self.assertRaises(Exception):
+            broken()
+        expected_output_in_file = 'raise Exception("This is a log message")'
         self.logfile_meets_expectation(expected_output_in_file)
 
     def test_shell_handler(self):
@@ -39,12 +48,15 @@ class TestPyttmanLogger(PyttmanInternalBaseTestCase):
         with open(self.log_file_name, "r") as file:
             match = False
             for line in file.readlines():
-                if line.strip() == expected_output_in_file:
+                if expected_output_in_file in line:
                     match = True
+                    break
         self.assertTrue(match)
 
     def cleanup(self):
-        print(Path().cwd())
+        if not self.cleanup_after:
+            return
+
         for logfile in Path().cwd().parent.parent.glob("*.log"):
             try:
                 os.remove(logfile)
